+++ resources/js/stores/settings.js
@@ -1,6 +1,7 @@
 import { defineStore } from 'pinia';
 import _ from 'lodash';
 import { useApi } from '../composables/useApi.js';
+import { setToastLifetime } from '@/composables/useToast';
 
 export const useSettingsStore = defineStore('settings', {
   state: () => {
@@ -19,6 +20,8 @@ export const useSettingsStore = defineStore('settings', {
 
       const response = await api.call('settings');
       this.settings = response.data.data;
+
+      setToastLifetime(this.settings.toast_lifetime);
     }
   }
 });